transition to basegraph 0.4; metrics are now templates
